+++ src/main/java/com/cafe/cafeapp/service/EmployeeService.java
@@ -3,9 +3,11 @@ package com.cafe.cafeapp.service;
 import com.cafe.cafeapp.dto.EmployeeRequestDto;
 import com.cafe.cafeapp.dto.EmployeeResponseDto;
 import com.cafe.cafeapp.exception.AlreadyExistsException;
+import com.cafe.cafeapp.exception.BusinessException;
 import com.cafe.cafeapp.exception.NotFoundException;
 import com.cafe.cafeapp.mapper.EmployeeMapper;
 import com.cafe.cafeapp.model.Employee;
+import com.cafe.cafeapp.model.Order;
 import com.cafe.cafeapp.repository.EmployeeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -65,6 +67,13 @@ public class EmployeeService {
     public void delete(Long id) {
         if (!employeeRepository.existsById(id)) {
             throw new NotFoundException(id);
+        }
+
+        Employee employee = employeeRepository.findById(id).get();
+        List<Order> orders = employee.getOrders();
+
+        if (!orders.isEmpty()) {
+            throw new BusinessException(id);
         }
 
         employeeRepository.deleteById(id);
